contacts controller separated
