--- controllers/contact-controller.js
@@ -1,103 +1,0 @@
-const Contact = require("../models/contact");
-
-// const contactsService = require("../models/contactsOld");
-const { HttpError } = require("../helpers");
-const { contactAddSchema, contactUpdateFavoriteSchema } = require("../schemas");
-
-const listContacts = async (req, res, next) => {
-  try {
-    const result = await Contact.find({}, "name email phone favorite");
-    res.json(result);
-  } catch (error) {
-    next(error);
-  }
-};
-
-const addContact = async (req, res, next) => {
-  try {
-    const { error } = contactAddSchema.validate(req.body);
-    if (error) {
-      throw HttpError(400, error.message);
-    }
-    const result = await Contact.create(req.body);
-    res.status(201).json(result);
-  } catch (error) {
-    next(error);
-  }
-};
-
-const getContactById = async (req, res, next) => {
-  try {
-    console.log(req.params);
-    const { contactId } = req.params;
-    const result = await Contact.findById(contactId);
-    if (!result) {
-      throw HttpError(404, `Contact not found with id: ${contactId}`);
-    }
-    res.json(result);
-  } catch (error) {
-    next(error);
-  }
-};
-
-const updateContact = async (req, res, next) => {
-  try {
-    const { error } = contactAddSchema.validate(req.body);
-
-    if (error) {
-      throw HttpError(400, error.message);
-    }
-    const { contactId } = req.params;
-    const result = await Contact.findByIdAndUpdate(contactId, req.body, { new: true });
-
-    if (!result) {
-      throw HttpError(404, `Contact with id: ${contactId} not found`);
-    }
-
-    res.status(200).json(result);
-  } catch (error) {
-    next(error);
-  }
-};
-
-const updateFavorite = async (req, res, next) => {
-  try {
-    const { error } = contactUpdateFavoriteSchema.validate(req.body);
-
-    if (error) {
-      throw HttpError(400, error.message);
-    }
-    const { contactId } = req.params;
-    const result = await Contact.findByIdAndUpdate(contactId, req.body, { new: true });
-
-    if (!result) {
-      throw HttpError(404, `Contact with id: ${contactId} not found`);
-    }
-
-    res.status(200).json(result);
-  } catch (error) {
-    next(error);
-  }
-};
-
-const removeContact = async (req, res, next) => {
-  try {
-    const { contactId } = req.params;
-    const result = await Contact.findByIdAndDelete(contactId);
-    if (!result) {
-      throw HttpError(404, `Contact not found with id: ${contactId}`);
-    }
-    res.json({ message: "Contact deleted successfully" });
-  } catch (error) {
-    next(error);
-  }
-};
-
-module.exports = {
-  listContacts,
-  addContact,
-  getContactById,
-  updateContact,
-  updateFavorite,
-  removeContact,
-};
