+++ controllers/contacts/getContactById.js
@@ -1,0 +1,19 @@
+const Contact = require("../../models/contact");
+
+const { HttpError } = require("../../helpers");
+const { contactAddSchema } = require("../../schemas");
+
+const addContact = async (req, res, next) => {
+  try {
+    const { error } = contactAddSchema.validate(req.body);
+    if (error) {
+      throw HttpError(400, error.message);
+    }
+    const result = await Contact.create(req.body);
+    res.status(201).json(result);
+  } catch (error) {
+    next(error);
+  }
+};
+
+module.exports = addContact;
